refactor: 시설 타입별로 폴더 정리
