+++ gogildong/src/main/java/com/efub/gogildong/reports/dto/request/classroom/ClassroomReportRequest.java
@@ -1,4 +1,4 @@
-package com.efub.gogildong.reports.dto.request;
+package com.efub.gogildong.reports.dto.request.classroom;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class RestRoomReportRequest extends NewRestRoomReportRequest {
+public class ClassroomReportRequest extends NewClassroomReportRequest {
     private Long facilityId;
 }
