+++ gogildong/src/main/java/com/efub/gogildong/reports/dto/request/restroom/RestRoomReportRequest.java
@@ -1,4 +1,4 @@
-package com.efub.gogildong.reports.dto.request;
+package com.efub.gogildong.reports.dto.request.restroom;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ClassroomReportRequest extends NewClassroomReportRequest {
+public class RestRoomReportRequest extends NewRestRoomReportRequest {
     private Long facilityId;
 }
