+++ gogildong/src/main/java/com/efub/gogildong/reports/dto/response/restroom/RestRoomReportResponse.java
@@ -1,4 +1,4 @@
-package com.efub.gogildong.reports.dto.response;
+package com.efub.gogildong.reports.dto.response.restroom;
 
 import com.efub.gogildong.facility.domain.DoorType;
 import com.efub.gogildong.facility.domain.GenderType;
