+++ gogildong/src/main/java/com/efub/gogildong/reports/service/ReportService.java
@@ -9,7 +9,17 @@ import com.efub.gogildong.global.exception.GoGildongException;
 import com.efub.gogildong.global.util.EntityFinder;
 import com.efub.gogildong.reports.domain.*;
 import com.efub.gogildong.reports.dto.request.*;
+import com.efub.gogildong.reports.dto.request.classroom.ClassroomReportRequest;
+import com.efub.gogildong.reports.dto.request.classroom.NewClassroomReportRequest;
+import com.efub.gogildong.reports.dto.request.elevator.ElevatorReportRequest;
+import com.efub.gogildong.reports.dto.request.elevator.NewElevatorReportRequest;
+import com.efub.gogildong.reports.dto.request.restroom.NewRestRoomReportRequest;
+import com.efub.gogildong.reports.dto.request.restroom.RestRoomReportRequest;
 import com.efub.gogildong.reports.dto.response.*;
+import com.efub.gogildong.reports.dto.response.classroom.ClassroomAggregateStat;
+import com.efub.gogildong.reports.dto.response.elevator.ElevatorAggregateStat;
+import com.efub.gogildong.reports.dto.response.restroom.RestRoomAggregateStat;
+import com.efub.gogildong.reports.dto.response.restroom.RestRoomReportResponse;
 import com.efub.gogildong.reports.dto.summary.ReportFlagSummary;
 import com.efub.gogildong.reports.dto.summary.ReportSummary;
 import com.efub.gogildong.reports.repository.*;
